modify athentication
