+++ controllers/auth/login.js
@@ -13,10 +13,6 @@ const login = async (req, res) => {
       return res.status(400).json({Status:0, Message: 'Invalid credentials' });
     }
 
-    // Check if the user is verified
-    if (!user.isVerified) {
-      return res.status(400).json({Status:0, Message: 'Email is not verified' });
-    }
 
     // Check if the password matches
     const isMatch = await bcrypt.compare(password, user.password);
@@ -26,6 +22,13 @@ const login = async (req, res) => {
 
     // Generate JWT token
     const token = generateToken({ userId: user._id, email: user.email });
+    
+    res.cookie('token', token, {
+      httpOnly: true, // Prevents client-side JavaScript from accessing the cookie
+      maxAge: 3 * 24 * 60 * 60 * 1000, // 3 days expiration in milliseconds
+      secure: process.env.NODE_ENV === 'production', // Ensures cookies are only sent over HTTPS in production
+      sameSite: 'strict', // Prevent CSRF attacks
+    });
 
     return res.status(200).json({Status:1, Message: 'Login successful', token });
   } catch (error) {
