+++ controllers/auth/register.js
@@ -1,11 +1,10 @@
 const bcrypt = require('bcryptjs');
 const User = require('../../models/User');
-const { generateToken } = require('../../config/configJWT');
 const { sendOTP } = require('../../config/configSMTP');
 const crypto = require('crypto');
 
 // Temporary storage for OTP (This can be replaced with Redis or DB for production)
-let otpStorage = {};
+const otpStorage = require('../../utils/auth/otpStorage');  //Import otpStorage
 
 const register = async (req, res) => {
     const { name, username, email, password, phoneNumber } = req.body;
@@ -64,67 +63,5 @@ const register = async (req, res) => {
   };
   
 
-// Function to handle OTP resending
-const resendOTP = async ( req, res) => {
-    const { email } = req.body;
-    try {
-        // Find the user by their ID (you can also find by email if you prefer)
-        const user = await User.findOne({ email });
-        if (!user) {
-            return res.status(400).json({Status:0, Message: 'User Not Found' });
-        }
 
-        // Generate a new OTP
-        const otp =  crypto.randomInt(100000, 999999);
-
-        // Store the new OTP in the user's record (you should have an OTP field and OTP expiry logic)
-        otpStorage[email] = { otp, expiresIn: Date.now() + 10 * 60 * 1000 };
-
-        // Send the OTP to the user's email
-        await sendOTP(email, otp);
-
-        return res.status(200).json({
-            Status:1,  
-            Message: 'OTP resent successfully'
-          });
-
-    } catch (error) {
-        return res.status(500).json({Status:0, Message: 'Something went wrong', error });
-    }
-}
-
-const verifyOtp = async (req, res) => {
-  const { email, otp } = req.body;
-
-  try {
-    const storedOtpData = otpStorage[email];
-
-    // Check if OTP exists and has not expired
-    if (!storedOtpData || storedOtpData.expiresIn < Date.now()) {
-      return res.status(400).json({Status:0, Message: 'OTP has expired or invalid' });
-    }
-
-    // Check if OTP is correct
-    if (parseInt(otp) !== storedOtpData.otp) {
-      return res.status(400).json({Status:0, Message: 'Invalid OTP' });
-    }
-
-    // Update user to verified
-    const user = await User.findOne({ email });
-    if (!user) {
-      return res.status(404).json({ Status:0,Message: 'User not found' });
-    }
-
-    // Generate JWT token after successful OTP verification
-    const token = generateToken({ userId: user._id, email: user.email });
-
-    user.isVerified = true;
-    await user.save();
-
-    return res.status(200).json({Status:1, Message: 'User verified successfully', token });
-  } catch (error) {
-    return res.status(500).json({Status:0, Message: 'Something went wrong', error });
-  }
-};
-
-module.exports = { register, resendOTP, verifyOtp };
+module.exports = { register };
