+++ routes/auth.js
@@ -2,7 +2,9 @@
 
 const express = require('express');
 const router = express.Router();
-const { register, verifyOtp, resendOTP} = require('../controllers/auth/register');
+const { register } = require('../controllers/auth/register');
+const { resendOTP } = require('../utils/auth/resentOTP');
+const { verifyOtp } = require('../utils/auth/verifyOTP');
 const { login } = require('../controllers/auth/login');
 
 router.post('/register', register);
