+++ utils/auth/otpStorage.js
@@ -1,0 +1,7 @@
+// otpStorage.js
+
+// Temporary storage for OTPs (in-memory)
+const otpStorage = {};
+
+// Export the otpStorage object so it can be used in other files
+module.exports = otpStorage;
